% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_ghostscript.R
\name{detect_ghostscript}
\alias{detect_ghostscript}
\title{Detect ghostscript installation}
\usage{
detect_ghostscript(quiet = F)
}
\arguments{
\item{quiet}{Do not show messages}
}
\value{
path to ghostscript.
}
\description{
Detect ghostscript installation
}
