% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghostpdf.R
\docType{package}
\name{ghostpdf}
\alias{ghostpdf}
\title{ghostpdf: A package for combining multiple pdfs}
\description{
Provides a high level interface for some of ghostscript's functions allowing the user to
convert between file formats, combine multiple pdfs and eps, etc
}
\section{ghostpdf functions}{

pdf_combine
}

