% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_pdfs.R
\name{pdf_combine}
\alias{pdf_combine}
\title{PDF Combine}
\usage{
pdf_combine(
  pdf_files_to_paste = list.files(pattern = "\\\\.pdf"),
  output_file = "combined_pdfs.pdf",
  auto_rotate_pages = c("None", "All", "PageByPage"),
  gs_path = NULL,
  quiet = F,
  cleanup = F
)
}
\arguments{
\item{pdf_files_to_paste}{vector with pdf files to combine in one unique pdf or
a character indicating which pdf's (such as *.pdf)}

\item{output_file}{name of the output pdf that combines all pdfs}

\item{auto_rotate_pages}{automatically rotate pages when pasting (default = \code{FALSE})}

\item{gs_path}{path to Ghostscript installation}

\item{quiet}{Don't output messages}

\item{cleanup}{Delete the files after merge? Default is \code{F}}
}
\description{
Combine multiple pdf's in one unique file
}
\details{
Needs Ghostscript installation to work
}
\examples{
for (i in 0:9){
  pdf(paste0("Example",i,".pdf"), width = 6, height = 4)
  plot(rnorm(10), rnorm(10), main = paste("Example",i,".pdf"))
  dev.off()
}

pdf_combine(list.files(pattern = "\\\\.pdf"), output_file = "test.pdf", cleanup = TRUE)
file.remove("test.pdf")
}
