% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_decrypt.R
\name{pdf_decrypt}
\alias{pdf_decrypt}
\title{PDF Decrypt}
\usage{
pdf_decrypt(
  pdf_file,
  password,
  output_file = pdf_file,
  quiet = T,
  gs_path = NULL,
  additional_flags = ""
)
}
\arguments{
\item{pdf_file}{pdf file from which to count the number of pages}

\item{password}{Defines that the document be encrypted with the specified owner password.
Owner password sets the password to edit the document.}

\item{output_file}{output pdf file of reduced size}

\item{quiet}{Don't output messages}

\item{gs_path}{path to Ghostscript installation}

\item{additional_flags}{Additional string of bash flags to pass to ghostscript}
}
\description{
Decrypts a password protected pdf file
}
\details{
Needs Ghostscript installation to work.
}
\examples{
pdf("Example.pdf", width = 6, height = 4)
plot(rnorm(10), rnorm(10), main = "Encrypted example")
dev.off()

#Shrink file
pdf_encrypt("Example.pdf", password = "mypassword", output_file = "pdf_encrypted.pdf")
pdf_decrypt("Example.pdf", password = "mypassword", output_file = "pdf_decrypted.pdf")
file.remove(c("Example.pdf","pdf_encrypted.pdf","pdf_decrypted.pdf"))
}
