% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_encrypt.R
\name{pdf_encrypt}
\alias{pdf_encrypt}
\title{PDF Encrypt}
\usage{
pdf_encrypt(
  pdf_file,
  password,
  user_password = password,
  output_file = pdf_file,
  quiet = T,
  gs_path = NULL,
  owner_password = password,
  permissions = NULL,
  encryption_revision = 3,
  key_length = 128,
  additional_flags = ""
)
}
\arguments{
\item{pdf_file}{pdf file from which to count the number of pages}

\item{password}{Set same password for owner and user. Ignored if \code{owner_password} and
\code{user_password} are set.}

\item{user_password}{Defines the user password for opening the document but cannot edit.}

\item{output_file}{output pdf file of reduced size}

\item{quiet}{Don't output messages}

\item{gs_path}{path to Ghostscript installation}

\item{owner_password}{Defines that the document be encrypted with the specified owner password.
Owner password sets the password to edit the document.}

\item{permissions}{Defines the PDF permissions flag field. Negative values are allowed to
represent unsigned integers with the highest bit set. See the PDF Reference manual
for the meaning of the flag bits.}

\item{encryption_revision}{Defines the encryption method revision number - either 2 or 3.}

\item{key_length}{Defines the length (in bits) of the encryption key. Must be a multiple of
8 in the interval (40, 128). If the length isn't 40, \code{encryption_revision} must be 3.}

\item{additional_flags}{Additional string of bash flags to pass to ghostscript}
}
\description{
Encrypts a pdf file
}
\details{
Needs Ghostscript installation to work.
}
\examples{
pdf("Example.pdf", width = 6, height = 4)
plot(rnorm(10), rnorm(10), main = "Encrypted example")
dev.off()

#Shrink file
pdf_encrypt("Example.pdf", password = "mypassword")
file.remove("Example.pdf")
}
