% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_extract.R
\name{pdf_extract}
\alias{pdf_extract}
\title{PDF Extract}
\usage{
pdf_extract(
  pdf_file,
  first_page = NULL,
  last_page = NULL,
  page_sequence = NULL,
  output_file = "output.pdf",
  preserve_annotations = TRUE,
  gs_path = NULL,
  quiet = T,
  auto_rotate_pages,
  additional_flags = ""
)
}
\arguments{
\item{pdf_file}{pdf file from which extract pages}

\item{first_page}{first page to extract}

\item{last_page}{last page to extract}

\item{page_sequence}{page numbers to extract if \code{first_page} and \code{last_page} are
not specified you can create a page_sequence of numbers of the pdf pages to extract}

\item{output_file}{name of the output pdf that combines all pdfs}

\item{preserve_annotations}{boolean flag to try and preserve annotations in the pdfs (default TRUE)}

\item{gs_path}{path to Ghostscript installation}

\item{quiet}{Don't output messages}

\item{auto_rotate_pages}{automatically rotate pages when pasting (default = \code{"None"})}

\item{additional_flags}{Additional string of bash flags to pass to ghostscript}
}
\description{
Extract pages from a pdf
}
\details{
You can either extract a continuous sequence of pages specifying  \code{first_page}
and \code{last_page} or a discontinuous sequence by using \code{page_sequence}.
Needs Ghostscript installation to work.
}
\examples{
for (i in 1:9){
  pdf(paste0("Example",i,".pdf"), width = 6, height = 4)
  plot(rnorm(10), rnorm(10), main = paste("Page",i,".pdf"))
  dev.off()
}

pdf_merge(list.files(pattern = "\\\\.pdf"), output_file = "test.pdf", cleanup = TRUE)

#Get pages 4 to 7
pdf_extract("test.pdf", 4, 7, output_file = "example_1.pdf")
file.remove("example_1.pdf")

#Get pages 1, 5, 7
pdf_extract("test.pdf", page_sequence = c(1,5,7), output_file = "example_2.pdf")
file.remove("example_2.pdf")

#Get pages 5 till the end
pdf_extract("test.pdf", first_page = 5, output_file = "example_3.pdf")
file.remove("example_3.pdf")

file.remove("test.pdf")
}
