% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_merge.R
\name{pdf_merge}
\alias{pdf_merge}
\title{PDF Combine}
\usage{
pdf_merge(
  pdf_files_to_paste = list.files(pattern = "\\\\.pdf"),
  output_file = "output.pdf",
  auto_rotate_pages = c("None", "All", "PageByPage"),
  pdf_settings = c("default", "ebook", "screen", "printer", "prepress"),
  gs_path = NULL,
  quiet = TRUE,
  cleanup = FALSE,
  preserve_annotations = TRUE,
  detect_duplicate_images = FALSE,
  fast_web_view = FALSE,
  max_inline_image_size = NULL,
  additional_flags = c("-dAutoFilterColorImages=false", "-dAutoFilterGrayImages=false",
    "-dColorImageFilter=/FlateEncode", "-dGrayImageFilter=/FlateEncode",
    "-dDownsampleMonoImages=false", "-dDownsampleGrayImages=false")
)
}
\arguments{
\item{pdf_files_to_paste}{vector with pdf files to combine in one unique pdf or
a character indicating which pdf's (such as *.pdf)}

\item{output_file}{name of the output pdf that combines all pdfs}

\item{auto_rotate_pages}{automatically rotate pages when pasting (default = \code{"None"})}

\item{pdf_settings}{Presets the "distiller parameters" to one of four predefined settings:
"eBook" \code{"ebook"}, "Screen Optimized" \code{"screen"},
"Print Optimized" \code{"printer"}, "Prepress Optimized" \code{"prepress"} or
\code{"default"}}

\item{gs_path}{path to Ghostscript installation}

\item{quiet}{Don't output messages}

\item{cleanup}{Delete the files after merge? Default is \code{F}}

\item{preserve_annotations}{boolean flag to try and preserve annotations in the pdfs (default TRUE)}

\item{detect_duplicate_images}{boolean that attempts to detect the same image file
and keep only one version of the image to reduce pdf size.}

\item{fast_web_view}{boolean that reorders the output PDF file to conform to the Adobe
'linearised' PDF specification}

\item{max_inline_image_size}{Specifies the maximum size of an inline image, in bytes.
For images larger than this size, pdfwrite will create an XObject instead of embedding the
image into the context stream. The default value is 4000. Note that redundant inline images
must be embedded each time they occur in the document, while multiple references can be made
to a single XObject image. Therefore it may be advantageous to set a small or zero value if
the source document is expected to contain multiple identical images, reducing the size of
the generated PDF.}

\item{additional_flags}{Additional string of bash flags to pass to ghostscript}
}
\description{
Combine multiple pdf's in one unique file
}
\details{
Needs Ghostscript installation to work
}
\examples{
for (i in 1:9){
  pdf(paste0("Example",i,".pdf"), width = 6, height = 4)
  plot(rnorm(10), rnorm(10), main = paste("Example",i,".pdf"))
  dev.off()
}

#Combine the pdf's
pdf_merge(list.files(pattern = "Example.*\\\\.pdf"), output_file = "merge_1.pdf",
cleanup = TRUE)
file.remove("merge_1.pdf")
}
