% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_num_pages.R
\name{pdf_num_pages}
\alias{pdf_num_pages}
\title{PDF Number of Pages}
\usage{
pdf_num_pages(pdf_file, gs_path = NULL, quiet = T, additional_flags = "")
}
\arguments{
\item{pdf_file}{pdf file from which to count the number of pages}

\item{gs_path}{path to Ghostscript installation}

\item{quiet}{Don't output messages}

\item{additional_flags}{Additional string of bash flags to pass to ghostscript}
}
\description{
Get the number of pages on a pdf file
}
\details{
Needs Ghostscript installation to work.
}
\examples{
for (i in 0:9){
  pdf(paste0("Example",i,".pdf"), width = 6, height = 4)
  plot(rnorm(10), rnorm(10), main = paste("Example",i,".pdf"))
  dev.off()
}

pdf_merge(list.files(pattern = "\\\\.pdf"), output_file = "test.pdf", cleanup = TRUE)
pdf_num_pages("test.pdf")
file.remove("test.pdf")
}
