% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghostpdf.R
\docType{package}
\name{ghostpdf}
\alias{ghostpdf}
\title{ghostpdf: A package for combining multiple pdfs}
\description{
The ghostpdf package provides a high level interface for some of ghostscript's
functions.
}
\section{ghostpdf functions}{

paste_pdfs
}

